webImage/requirements.txt
asgiref==3.8.1
awscli==1.38.9
boto3==1.37.6
botocore==1.37.9
colorama==0.4.6
Django==4.2.19
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.16
git-filter-repo==2.47.0
jmespath==1.0.1
load-dotenv==0.1.0
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rsa==4.7.2
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
dj_database_url
torch 
torchvision
faiss-cpu 
numpy 
requests 
pillow 
tqdm